test(search): split dot-directory tests into separate file (FR-006)
src/tools/search.test.ts was already over the 500-line cap; the new
#2212 tests pushed it further over. Split them into
search-dotdir-security.test.ts to satisfy the diff-scoped ratchet.
